@@ -1,7 +1,16 @@
 import { app, BrowserWindow } from 'electron';
 import installExtension, { REACT_DEVELOPER_TOOLS } from 'electron-devtools-installer';
 import { enableLiveReload } from 'electron-compile';
-const path = require('path')
+import * as path from 'path'
+import { register } from 'tsconfig-paths'
+
+// tslint:disable-next-line:no-var-requires no-require-imports
+const config = require('../tsconfig.json')
+
+register({
+  baseUrl: config.compilerOptions.baseUrl,
+  paths: config.compilerOptions.paths,
+})
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
