@@ -1,2 +1,5 @@
-export * from './CreateWallet'
-export * from './Dashboard'
+export { CreateWallet } from './CreateWallet'
+export { Dashboard } from './Dashboard'
+export { AppSettings } from './AppSettings'
+export { WalletList } from './WalletList'
+export { App } from './Main'
